add promo tests to marketplace e2e tests so they can be run after the marketplaceis installed, but before premium features are installed
