+++ wdio.conf.ts
@@ -113,7 +113,6 @@ export const config: Options.Testrunner = {
   // Patterns to exclude. (all executed without parallelism before these tests, see wdio.conf.tracking.ts)
   exclude: [
     './tests/e2e/update.e2e.ts',
-    './tests/e2e/matomo-reporting.promos.e2e.ts',
     './tests/e2e/mwp-admin.marketplace.e2e.ts',
     './tests/e2e/tracking.e2e.ts',
     './tests/e2e/tracking.ecommerce.e2e.ts',
